requirements.txt
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
greenlet==1.1.0
idna==2.10
pycodestyle==2.7.0
PyMySQL==1.0.2
python-dotenv==0.17.1
requests==2.25.1
SQLAlchemy==1.4.16
toml==0.10.2
urllib3==1.26.5
